Memindahkan LadderMovement ke PlayerController
