--- Assets/Scripts/LadderMovement.cs
@@ -1,58 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class LadderMovement : MonoBehaviour
-{
-    private float vertical;
-    private float speed = 8f;
-    private bool isLadder;
-    private bool isClimbing;
-
-    private float defaultGravityScale ;
-    [SerializeField] private Rigidbody2D rb;
-
-    void Start(){
-        defaultGravityScale =  rb.gravityScale;
-    }
-
-    void Update()
-    {
-        vertical = Input.GetAxisRaw("Vertical");
-
-        if (isLadder && Mathf.Abs(vertical) > 0f)
-        {
-            isClimbing = true;
-        }
-    }
-
-    private void FixedUpdate()
-    {
-        if (isClimbing)
-        {
-            rb.gravityScale = 0f;
-            rb.velocity = new Vector2(rb.velocity.x, vertical * speed);
-        }
-        else
-        {
-            rb.gravityScale =defaultGravityScale;
-        }
-    }
-
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.CompareTag("Ladder"))
-        {
-            isLadder = true;
-        }
-    }
-
-    private void OnTriggerExit2D(Collider2D collision)
-    {
-        if (collision.CompareTag("Ladder"))
-        {
-            isLadder = false;
-            isClimbing = false;
-        }
-    }
-}
