--- Assets/Scripts/LadderMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3e6cf6b73016967f8a677a98bc083733
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
